refactor(shared): extract EventEmitter, safe_json_loads, normalize_text
Council-style plugins share small response parsing and status emission utilities. Move them into shared modules to reduce duplicate source.
