+++ src/owui_ext/tools/magi_decision_support.py
@@ -27,6 +27,8 @@ from fastapi import Request
 from pydantic import BaseModel, Field
 
 from owui_ext.shared.async_utils import maybe_await
+from owui_ext.shared.event_emitter import EventEmitter
+from owui_ext.shared.parsing import normalize_text, safe_json_loads
 from owui_ext.shared.prompt_utils import (
     _append_tool_server_prompts,
     merge_prompt_sections,
@@ -49,53 +51,6 @@ log = logging.getLogger(__name__)
 
 
 WEB_TOOL_NAMES = {"search_web", "fetch_url"}
-
-
-class EventEmitter:
-    def __init__(self, event_emitter: Callable[[dict], Any] = None):  # type: ignore
-        self.event_emitter = event_emitter
-
-    async def emit(
-        self, description: str = "Unknown state", status: str = "in_progress", done: bool = False
-    ) -> None:
-        if self.event_emitter:
-            await self.event_emitter(
-                {
-                    "type": "status",
-                    "data": {
-                        "status": status,
-                        "description": description,
-                        "done": done,
-                    },
-                }
-            )
-
-
-def safe_json_loads(text: str) -> Optional[dict]:
-    if not text:
-        return None
-    text = text.strip()
-    try:
-        return json.loads(text)
-    except Exception:
-        pass
-
-    start = text.find("{")
-    end = text.rfind("}")
-    if start >= 0 and end > start:
-        candidate = text[start : end + 1]
-        try:
-            return json.loads(candidate)
-        except Exception:
-            return None
-
-    return None
-
-
-def normalize_text(value: Optional[str]) -> str:
-    if not value:
-        return ""
-    return str(value).strip()
 
 
 async def resolve_terminal_id_from_request_and_metadata(
